fix (arredondamento/infra): Resolve bug que arredondava o saldo na entrada da requisição ao sistema pela api, além de melhorar a fluidez e autoexplicação do código.

Contexto
A aplicação arredondava 21.99999 para 22 no momento da entrada do saldo no sistema pela API, devido à configuração suportar apenas 19 inteiros e 4 casas decimais. Assim, a conversão ocorria de forma padrão.

O sistema também apresentava as classes de serviço de forma complexa e verbosa.

Método de resolução
Para o arredondamento, foi desenvolvida uma validação do formato do saldo na API, que agora rejeita formatos inválidos.

Também foi criado o VO (Value Object) Money, com o intuito de padronizar os métodos que lidam com valores monetários em uma única classe, visando limitar ações desnecessárias, manter um padrão e permitir a existência apenas de objetos válidos.

Além disso, para reduzir a complexidade dentro das classes de serviço, foram desenvolvidos factory objects para as entidades, além de pequenas validações que impedem a criação de instâncias inválidas.

+++ src/main/java/com/finance/wallet/v12/dto/response/TransactionResponseDTO.java
@@ -22,7 +22,7 @@ public record TransactionResponseDTO(
                 transaction.getWalletReceiver().getId(),
                 transaction.getWalletSender() != null ?
                         transaction.getWalletSender().getId() : null,
-                transaction.getAmount(),
+                transaction.getAmount().toBigDecimal(),
                 transaction.getCreatedAt(),
                 transaction.getOperationType()
         );
